Replace Flex component
