+++ src/components/molecules/BlogPostTitle/BlogPostTitle.module.scss
@@ -1,0 +1,5 @@
+.root {
+  @apply flex flex-col justify-start items-start w-full;
+
+  max-width: 700px;
+}
