+++ src/components/molecules/MainLayout/MainLayout.module.scss
@@ -1,0 +1,5 @@
+.article {
+  @apply flex-col justify-center items-start mx-auto mb-16 w-full;
+
+  max-width: 700px;
+}
